requirements.txt
# pip install GitPython
GitPython==3.1.20